Add download methods to commonsdownload
Some of the downloading logic was nested in the main method,
which is no good if we want to reuse them in a different context.

Move the logic to two methods, one for downloading from a text list,
one from a Python list of files.
Call these methods in the main.