Add method to download a full size image
We still want to be able to retrieve a full size version of a file.

Add method get_full_size_file which downloads the full size version of a given file.
(it makes use of the method to build the full size URL, introduced in de99534).
Add relevant unit test with its data file.