Make download_file work with width bigger than source
"When I specify a width for a file that is larger than the file width,
I want CommonsDownloader to get the largest thumb possible for this file
instead of getting an error.”

This finishes the implementation of the above user story, building
on all commits from 2046595 to 0c8732f.

When the user requests a width larger than source, get_thumbnail_of_file
throws an Exception, which is caught. Then get_full_size_file is called
instead and downloads the file in its full size.

Edit unit tests to reflect this new behaviour.
Remove test_download_file_with_size_higher_than_possible
as it does not have a special behaviour anymore.
Add general test case with full size file.

Resolves #2.