Add caching to CommonsDownloader
As a user, I do not want CommonsDownloader to download
again all the files I have already downloaded.

This implements the above user story by using a local cache,
as a file stored in the output directory, containing the file
names and the widths downloaded (in the same format than the
input file).

CommonsDownloader only proceeds with downloading if the file
is not already present with the given width.

Resolves #4.