Extend error message matching for non existing full size image
In b180b258, we added to `get_exception_based_on_api_message` the
handling of `get_full_size_file` failing when the file does not exist.

The HTML that we were relying upon has been changed since 2013.

Let's just relax the matching, as this is poor design to begin with.